refactor: separate update and transfer scripts
