--- scripts/data.ts
@@ -1,7 +1,0 @@
-export const domains = [];
-
-export const owners = [];
-
-export const voltaPrivateKey = '';
-
-export const ewcPrivateKey = '';
