+++ scripts/transfer_domains.ts
@@ -1,9 +1,10 @@
 import { EWC_CHAIN_ID } from '@energyweb/credential-governance';
 import { Wallet } from 'ethers';
-import { transferDomain } from '../src/utils/domains';
-import { domains, ewcPrivateKey } from './data';
+import { transferDomain } from '../src/utils/transfer-domains';
 
-export const newOwner = '0x3451aaEDD3f25204D483aADCF060e344155DEB02';
+export const domains = ['energyweb.iam.ewc'];
+export const newOwner = '0x205801aAbf81B65B3b7C52041991994d6c2e9F9d';
+export const ewcPrivateKey = Wallet.createRandom().privateKey;
 
 (async function () {
   try {
@@ -16,8 +17,8 @@ export const newOwner = '0x3451aaEDD3f25204D483aADCF060e344155DEB02';
         dryRun: true,
       });
     }
-    console.log('Domains are transferred');
+    process.stdout.write('Domains are transferred\n');
   } catch (e) {
-    console.error('Error transferring domains:', e);
+    process.stderr.write(`Error transferring domains: ${e}\n`);
   }
 })();
